Documentation + code cleanup
