--- Labirynth/Material.h
@@ -1,36 +1,0 @@
-#pragma once
-
-#include<glew.h> //before glfw3 !!
-#include<glfw3.h>
-
-#include<glm.hpp>
-#include<vec2.hpp>
-#include<vec3.hpp>
-#include<vec4.hpp>
-#include<mat4x4.hpp>
-#include<gtc\type_ptr.hpp>
-
-#include"Shader.h"
-
-//used to symulate light  behavior on textures
-class Material {
-public:
-Material(
-	glm::vec3 ambient,
-	glm::vec3 diffuse,
-	glm::vec3 specular,
-	GLint diffuseTex,
-	GLint specularTex
-	);
-
-
-void sendToShader(Shader& program);
-
-private:
-	glm::vec3 ambient;
-	glm::vec3 diffuse;
-	glm::vec3 specular;
-	GLint diffuseTex;
-	GLint specularTex;
-
-};
