--- Labirynth/Model.h
@@ -1,33 +1,0 @@
-#pragma once
-
-#pragma once
-
-#include"Mesh.h"
-#include"Texture.h"
-#include"Shader.h"
-#include"Material.h"
-
-
-// connects meshes ,material and texture to create one model
-	
-class Model
-{
-private:
-	Material* material;
-	Texture* overrideTextureDiffuse;
-	Texture* overrideTextureSpecular;
-	std::vector<Mesh*> meshes;
-	glm::vec3 position;
-
-public:
-	Model(
-		glm::vec3 position,
-		Material* material,
-		Texture* orTexDif,
-		Texture* orTexSpec,
-		std::vector<Mesh*>& meshes);
-
-	~Model();
-
-	void render(Shader* shader);
-};
