--- Labirynth/Vertex.h
@@ -1,10 +1,0 @@
-#pragma once
-
-#include<glm.hpp>
-
-struct Vertex {
-	glm::vec3 position;
-	glm::vec3 color;
-	glm::vec2 texcoord;
-	glm::vec3 normal;
-};
