--- Labirynth/libs.h
@@ -1,25 +1,0 @@
-#pragma once
-#include<iostream>
-#include <fstream>
-#include <string>
-#include <vector>
-
-#include<glew.h> //before glfw3 !!
-#include<glfw3.h>
-
-#include<glm.hpp>
-#include<vec2.hpp>
-#include<vec3.hpp>
-#include<vec4.hpp>
-#include<mat4x4.hpp>
-#include<gtc\matrix_transform.hpp>
-#include<gtc\type_ptr.hpp>
-
-#include<SOIL2.h>
-#include "Mesh.h"
-#include "Camera.h"
-#include "Model.h"
-#include "Light.h"
-#include <time.h>
-#include <windows.h> 
-#include "MazeGenerator.h"
